Corrected mismatched types in process_generic_syscalls.
The prototype in syscall.h did not match the function definition
in syscall.c. This was only throwing errors on Paul's machine, for
some reason.

+++ kern/include/syscall.h
@@ -10,5 +10,5 @@
 intreg_t (syscall)(env_t* e, uintreg_t num, uintreg_t a1, uintreg_t a2,
                    uintreg_t a3, uintreg_t a4, uintreg_t a5);
 intreg_t syscall_async(env_t* e, syscall_req_t *syscall);
-intreg_t process_generic_syscalls(env_t* e, uint32_t max);
+intreg_t process_generic_syscalls(env_t* e, size_t max);
 #endif /* !ROS_KERN_SYSCALL_H */
